pyproject.toml
[project]
name = "face-detection-AI-model"
version = "0.1.0"
description = "capture the pic with name ,trained model and detect the face"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "streamlit>=1.45.0",
]
